Move cuda_event_pool function definitions to cpp file
